pipelines: add the createPipeline function + examples
The createPipeline function can be used to create derivations with
pipelines of sticker models. For instance, the following invocation
creates a tagging/parsing pipeline for Dutch:

nl-ud = createPipeline {
  name = "nl-ud";
  version = "20191003";
  models = with models; [
    nl-pos-ud
    nl-deps-ud-large
  ];
};

The output of the derivation contains the `sticker-tag-pipeline-nl-ud`
and `sticker-server-pipeline-nl-ud` utitilities. These wrap the
corresponding sticker programs, adding the models as arguments.