Pin the package set to a specific version of nixpkgs
This choice is not made lightly. This is not a general package set,
but a package set specifically provided to build sticker. The benefit
of pinning nixpkgs is twofold:

1. We are providing a version of the sticker set that we have tested.
2. Building a derivation from a particular revision of the package set
   becomes fully reproducible.

(There is still one wrinkle: the nightly Rust compiler for the Python
packages.)

Building against other versions of nixpkgs is still possible by
providing a `nixpkgs` argument.